requirements.txt
Django==4.1.7
djangorestframework==3.14.0
psycopg2-binary==2.9.5
python-dotenv==1.0.0
django-filter==22.1
drf-yasg==1.21.5
gunicorn==20.1.0